Workaround bug in tango.  :(
The bug is that after some number of runs (perhaps two, perhaps
even just one, not sure), it ends up treating the .rs file as
newer  than than the .md one, and so it does the reverse transformation,
which in this case is actually information destroying.

Arguably the better thing would be for tango to actually error on the
original input (two rust blocks in sequence with no intermediate content
at all) since the reverse transformation really is the one I want, and
the workaround I'm employing here is not *that* bad.